+++ src/textual/widget.py
@@ -3128,6 +3128,23 @@ class Widget(DOMNode):
         """
         self.app.capture_mouse(None)
 
+    def begin_capture_print(self, stdout: bool = True, stderr: bool = True) -> None:
+        """Capture text from print statements (or writes to stdout / stderr).
+
+        If printing is captured, the widget will be send an [events.Print][textual.events.Print] message.
+
+        Call [end_capture_print][textual.widget.Widget.end_capture_print] to disable print capture.
+
+        Args:
+            stdout: Capture stdout.
+            stderr: Capture stderr.
+        """
+        self.app.begin_capture_print(self, stdout=stdout, stderr=stderr)
+
+    def end_capture_print(self) -> None:
+        """End print capture (set with [capture_print][textual.widget.Widget.capture_print])."""
+        self.app.end_capture_print(self)
+
     def check_message_enabled(self, message: Message) -> bool:
         """Check if a given message is enabled (allowed to be sent).